requirements.txt
google-cloud-pubsub==2.13.7
GitPython==3.1.27
google-cloud-storage==2.5.0
google-cloud-bigquery==3.3.3
